+++ src/main/kotlin/io/github/freya022/botcommands/internal/commands/text/TextCommandComparator.kt
@@ -3,9 +3,9 @@ package io.github.freya022.botcommands.internal.commands.text
 import io.github.freya022.botcommands.api.core.BContext
 import io.github.freya022.botcommands.api.core.reflect.wrap
 import io.github.freya022.botcommands.api.core.service.getService
-import io.github.freya022.botcommands.api.parameters.ResolverContainer
 import io.github.freya022.botcommands.api.parameters.resolvers.TextParameterResolver
 import io.github.freya022.botcommands.internal.commands.text.autobuilder.metadata.TextFunctionMetadata
+import io.github.freya022.botcommands.internal.parameters.ResolverContainer
 import io.github.freya022.botcommands.internal.utils.ReflectionUtils.nonInstanceParameters
 import io.github.freya022.botcommands.internal.utils.shortSignatureNoSrc
 import io.github.oshai.kotlinlogging.KotlinLogging
